+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Fix bug when `keys` were set but setting cookies did not default to `signed`
+
 0.6.2 / 2016-11-12
 ==================